+++ apps/pooly/lib/pooly.ex
@@ -1,27 +1,25 @@
-defmodule Pooly.Application do
+defmodule Pooly do
   use Application
-
-  alias Pooly.{Server, Supervisor}
-  alias Pooly.SampleWorker
 
   def start(_type, _args) do
     pool_config = [mfa: {SampleWorker, :start_link, []}, size: 5]
     start_pool(pool_config)
   end
 
+  def start_pool(pool_config) do
+    Pooly.Supervisor.start_link(pool_config)
+  end
+
   def checkout do
-    Server.checkout()
+    Pooly.Server.checkout
   end
 
   def checkin(worker_pid) do
-    Server.checkin(worker_pid)
+    Pooly.Server.checkin(worker_pid)
   end
 
   def status do
-    Server.status()
+    Pooly.Server.status
   end
 
-  defp start_pool(pool_config) do
-    Supervisor.start_link(pool_config)
-  end
 end
